fix: sonar issues
