+++ frontend/src/app/pages/game-view/left-panel/actions-selector/actions-selector.component.ts
@@ -1,6 +1,5 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { WithAhSymbolsPipe } from 'shared/ui/pipes/with-ah-symbols.pipe';
-import { input } from '@angular/core';
 import { Actions } from 'shared/domain/action.model';
 
 @Component({
