+++ frontend/src/app/shared/domain/entities/location.model.ts
@@ -1,7 +1,6 @@
 ﻿import { type } from 'arktype';
 import { gameCard } from './card.model';
 import { locationId } from './id.model';
-import { GameEntity } from '../game-state';
 
 const _location = gameCard.and({
   id: locationId,
@@ -17,7 +16,3 @@ type _Location = typeof _location.infer;
 export interface Location extends _Location {}
 
 export const location: type<Location> = _location;
-
-export function isLocation(entity: GameEntity): entity is Location {
-  return entity.type === 'location';
-}
